requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Azure Storage SDK for blob storage
azure-storage-blob==12.19.0

# Environment variable management
python-dotenv==1.0.0

# WSGI HTTP Server for production
gunicorn==21.2.0
